use a new fork of the crdb driver
The original version of sqlboiler-crdb is unmaintained and did not
support CockroachDB versions after v.21, when support for dumping the
DDL for a database was dropped. Cockroach Labs replaced this with the
`SHOW CREATE-ALL TABLES`. The Metal Toolbox version of this driver
supports this new method but its generated code does not properly strip
quotes from the output of the command. Fleet Services forked this driver
again to address this issue. This change adapts the FleetDB repo for
this new driver by adding support for installing the driver directly,
and adapting the configuration and generated code to ease the friction
of changing CRDB or SQLBoiler versions.