alter event history table primary key
When I originally implemented this I envisioned using a single record
for each "workflow" managed by Condition-Orchestrator. For example, if
you did a firmware install you would get results only for the sum of the
install and the following inventory, instead of a record for each
condition we tried.

When I started looking at implementing the corresponding changes in
Condition-API and Condition-Orchestrator, I found that hooking into the
"combination" event was not as natural as providing a history record for
each condition being executed. However, adopting that strategy means
that I needed to change the table schema.